host-bmc: bypass getPDR when discovery the host
Ampere platforms use difference services to manage the host state and
do host soft off. host-bmc in pldm code checks the host state and call
GetPDRs when it detects the host is on. This can causes the confiction
with the GetPDRs commands in terminus manager because host can only
response for one GetPDRs command per time. This commit removes GetPDRs
process in host-bmc code.

Signed-off-by: ThuBaNguyen <thu@os.amperecomputing.com>
Change-Id: I9bb4dcbf0754001db4a71e62b7bd89d03fec8720