ampere: event: forward sensorEvent and pldmMessagePollEvent to Ampere app
The AmpereOne MPRo uses the numericSensorState event class of the
sensorEvent type to send the system errors/state code to BMC. These
error codes are encoded in the presentReading value. The error codes
are different between the sensors. They depend on the device which
contains the sensor. Using Ampere OEM PLDM code to handle these
platform events is not suitable because PLDM code is platform
independence.

This commit forwards sensorEvent and pldmMessagePollEvent to Ampere
PLDM event handler app. The app will handle the event data. It is
platform dependence and can be reused.

Tested:
1. Echo the numericSensorState event in the platform app.
2. Insert some sensor events in the MPRo.
3. The events should be logged in the journal log.

Signed-off-by: Thu Nguyen <thu@os.amperecomputing.com>