platform-mc: Handle Connectivity propertiesChanged signal
mctpd supports `.Connectivity` property under
au.com.CodeConstruct.MCTP.Endpoint interface of the endpoint. This
commit handles the propertiesChanged signal from this interface, and
updates the Availability of the MCTP Endpoint accordingly in the source
to enable or disable message sending/receiving via that endpoint of the
terminus.

When the discovery process first starts, it will only handle the
endpoints that have `Available` `.Connectivity`. It lets the
propertiesChanged signal trigger the adding of the endpoints when they
are back to Available.

On interfaceAdded signal, it assumes that mctpd only publishes available
endpoints to D-Bus, so it adds the endpoints to the terminus anyway.

Tested:
Enable unsafe-writable-connectivity option in PACKAGECONFIG of mctp
recipe. After PLDM discovers all the endpoints, write "Degraded" to
.Connectivity of one of the endpoint, and write it back to "Available"
to see how message is stopped from being sent/received via the endpoint.