platform-mc: PDR: Support Entity Aux Name PDRs
Section `Entity Auxiliary Names PDR` in DSP0248 details about the PDRs
response for the name of one PLDM entity. When the containerID is
`0x0000` this entity is `Overall system` or `top most containing
entity`. The name of this entity will be used as `Terminus name`.

Support parsing `Entity Auxiliary Names PDR` and find the `Terminus
name` if it is reponsed in the terminus' PDRs.

Signed-off-by: Thu Nguyen <thu@os.amperecomputing.com>