platform-mc: Initiate NaN for effecter D-Bus value
This commit initiates D-Bus Value property of a newly-created effecter
with NaN. This ensures a zero initial value to an effecter will still
result in a propertiesChanged signal, because without initializing, the
first value will be 0 and will avoid setting 0 again to D-Bus. That
signal missing will cause an issue in HostEffecterParser, as it relies
on the signals to handle setting values to the effecter on the other
terminus. The consequence is that, the next non-zero set to this
effecter will be mis-treated and have no effect.

Tested:

1. BMC boots up => PLDM service starts

Effecters with initial 0 value will have propertiesChanged signal
generated for Sensor.Value interface.

2. Set a non-zero value to the effecter

The value on D-Bus and of the effecter will be updated
correspondingly.

Change-Id: I8232f3963cae560f9962b895a1c26c1e01c00bbd
Signed-off-by: Chaul Ly <chaul@amperecomputing.com>